Split out command-line interface to new karabo_data.cli package
