+++ karabo_data/cli/__init__.py
@@ -1,0 +1,1 @@
+"""karabo_data command-line interfaces"""
