+++ setup.py
@@ -40,7 +40,7 @@ setup(name="karabo_data",
               "lsxfel = karabo_data.lsxfel:main",
               "karabo-bridge-serve-files = karabo_data.export:main",
               "karabo-data-validate = karabo_data.validation:main",
-              "karabo-data-make-virtual-cxi = karabo_data.write_cxi:main"
+              "karabo-data-make-virtual-cxi = karabo_data.cli.make_virtual_cxi:main"
           ],
       },
       install_requires=[
